fix: Refactor option handling for JWT strategy
