+++ packages/authentication/lib/jwt.js
@@ -1,10 +1,31 @@
 const { NotAuthenticated } = require('@feathersjs/errors');
-const BaseStrategy = require('./strategy');
 const SPLIT_HEADER = /(\S+)\s+(\S+)/;
 
-module.exports = class JWTStrategy extends BaseStrategy {
+module.exports = class JWTStrategy {
+  setAuthentication (auth) {
+    this.authentication = auth;
+  }
+
+  setApplication (app) {
+    this.app = app;
+  }
+
+  setName (name) {
+    this.name = name;
+  }
+
+  get configuration () {
+    const authConfig = this.authentication.configuration;
+    const config = authConfig[this.name];
+
+    return Object.assign({
+      entity: authConfig.entity,
+      service: authConfig.service
+    }, config);
+  }
+
   getEntity (id, params) {
-    const { service } = this.authentication.configuration;
+    const { service } = this.configuration;
     const entityService = this.app.service(service);
 
     if (!entityService) {
@@ -18,7 +39,7 @@ module.exports = class JWTStrategy extends BaseStrategy {
 
   authenticate (authentication, params) {
     const { accessToken, strategy } = authentication;
-    const { entity } = this.authentication.configuration;
+    const { entity } = this.configuration;
 
     if (!accessToken || (strategy && strategy !== this.name)) {
       return Promise.reject(new NotAuthenticated('Not authenticated'));
