--- packages/authentication/lib/strategy.js
@@ -1,17 +1,0 @@
-module.exports = class BaseStrategy {
-  setAuthentication (auth) {
-    this.authentication = auth;
-  }
-
-  setApplication (app) {
-    this.app = app;
-  }
-
-  setName (name) {
-    this.name = name;
-  }
-
-  get configuration () {
-    return this.authentication.configuration[this.name];
-  }
-};
